Split uniform_cube and uniform_sphere into separate problems
