+++ .github/workflows/uniform_sphere.yml
@@ -1,0 +1,39 @@
+name: uniform_sphere
+
+on: [pull_request]
+jobs:
+  uniform_sphere:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
+
+      - name: Get submodules
+        run: |
+          git submodule update --init
+          cd external/Microphysics
+          git fetch; git checkout development
+          cd ../amrex
+          git fetch; git checkout development
+          cd ../..
+
+      - name: Install dependencies
+        run: |
+          sudo apt-get update -y -qq
+          sudo apt-get -qq -y install curl g++>=9.3.0
+
+      - name: Compile uniform_sphere
+        run: |
+          cd Exec/gravity_tests/uniform_sphere
+          make USE_MPI=FALSE -j 2
+
+      - name: Run uniform_sphere
+        run: |
+          cd Exec/gravity_tests/uniform_sphere
+          ./convergence.sh &> sphere_convergence.out
+
+      - name: Compare to the benchmark
+        run: |
+          cd Exec/gravity_tests/uniform_sphere
+          diff sphere_convergence.out ci-benchmarks/sphere_convergence.out
