+++ Exec/gravity_tests/uniform_cube/ci-benchmarks/cube_convergence.out
@@ -1,0 +1,4 @@
+ncell = 16 error = 0.004276641412
+ncell = 32 error = 0.001071555867
+ncell = 64 error = 0.0002676508753
+Average convergence rate = 1.99932628187254717105
