+++ Exec/gravity_tests/uniform_cube/problem_initialize.H
@@ -2,10 +2,8 @@
 #define problem_initialize_H
 
 #include <prob_parameters.H>
-#include <eos.H>
 
 AMREX_INLINE
-void problem_initialize ()
-{
-}
+void problem_initialize () {}
+
 #endif
