+++ Exec/gravity_tests/uniform_sphere/Problem.H
@@ -1,0 +1,7 @@
+// Preprocessor directive for allowing us to do a post-initialization update.
+
+#ifndef DO_PROBLEM_POST_INIT
+#define DO_PROBLEM_POST_INIT
+#endif
+
+void problem_post_init();
