+++ Exec/gravity_tests/uniform_sphere/_prob_params
@@ -1,0 +1,5 @@
+ambient_dens       real     1.e-8_rt     y
+
+density            real     1.0_rt       y
+
+diameter           real     1.0_rt       y
