+++ Exec/gravity_tests/uniform_sphere/ci-benchmarks/sphere_convergence.out
@@ -1,0 +1,4 @@
+ncell = 16 error = 0.05274082586
+ncell = 32 error = 0.008316536152
+ncell = 64 error = 0.001270740323
+Average convergence rate = 2.53825936367487233285
