+++ Exec/gravity_tests/uniform_sphere/problem_initialize.H
@@ -1,0 +1,9 @@
+#ifndef problem_initialize_H
+#define problem_initialize_H
+
+#include <prob_parameters.H>
+
+AMREX_INLINE
+void problem_initialize () {}
+
+#endif
